// src/shared/api/models/postApiSaleBodyDataPaymentItemPaymentType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type PostApiSaleBodyDataPaymentItemPaymentType =
    (typeof PostApiSaleBodyDataPaymentItemPaymentType)[keyof typeof PostApiSaleBodyDataPaymentItemPaymentType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostApiSaleBodyDataPaymentItemPaymentType = {
    CASH: 'CASH',
    CARD: 'CARD',
    GIFT_CARD: 'GIFT_CARD',
    BUNDLE_OFFER: 'BUNDLE_OFFER',
    CHEQUE: 'CHEQUE',
    ECOMMERCE: 'ECOMMERCE',
    OTHER: 'OTHER',
    BANK_TRANSFER: 'BANK_TRANSFER',
    MOBILE_PAY: 'MOBILE_PAY',
    CUT_CARD: 'CUT_CARD',
    OUTSTANDING: 'OUTSTANDING',
    CASH_CREDIT: 'CASH_CREDIT',
    CARD_CREDIT: 'CARD_CREDIT',
    MOBILE_PAY_CREDIT: 'MOBILE_PAY_CREDIT',
    BANK_TRANSFER_CREDIT: 'BANK_TRANSFER_CREDIT',
    CHANGE: 'CHANGE',
    BONUS: 'BONUS',
} as const;